OnlineEditor/OnlineEditor/backend/UserServer/requirements.txt
aioredis==1.3.1
asgiref==3.7.2
astroid==2.4.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
brotlipy==0.7.0
certifi==2019.9.11
cffi==1.16.0
channels==3.0.4
channels-redis==3.3.0
chardet==3.0.4
colorama==0.4.4
constantly==23.10.4
cryptography==42.0.5
cssselect==1.1.0
daphne==3.0.2
Django==4.2.11
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-haystack==2.8.1
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.14.0
docxtpl==0.11.5
ecdsa==0.18.0
func-timeout==4.3.5
hiredis==2.3.2
hyperlink==21.0.0
idna==2.8
incremental==22.10.0
isort==5.13.2
jieba==0.42.1
Jinja2==2.10.1
lazy-object-proxy==1.4.3
lxml==5.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.8
#mysqlclient==2.0.3
olefile==0.46
#opencv-python==4.5.3.56
pillow==10.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pylint==2.6.0
pymongo==3.11.4
PyMySQL==1.1.0
pyOpenSSL==24.0.0
pyquery==1.4.0
PySocks==1.7.1
python-docx==0.8.7
python-jose==3.3.0
pytz==2019.2
qrcode==6.1
requests==2.22.0
rsa==4.9
service-identity==24.1.0
six==1.12.0
sqlparse==0.4.4
toml==0.10.2
twisted==24.3.0
txaio==23.1.1
typing-extensions==4.10.0
urllib3==1.25.6
Whoosh==2.7.4
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
yapf==0.28.0
zope.interface==6.2
